feat:move function about draw
