+++ Fusion/utils/visualize.py
@@ -1,8 +1,9 @@
 import cv2
 import numpy as np
+import matplotlib.pyplot as plt
+
 
 def vis(img, boxes, scores, cls_ids, conf=0.5, class_names=None):
-
     for i in range(len(boxes)):
         box = boxes[i]
         cls_id = int(cls_ids[i])
@@ -26,13 +27,44 @@ def vis(img, boxes, scores, cls_ids, conf=0.5, class_names=None):
         cv2.rectangle(
             img,
             (x0, y0 + 1),
-            (x0 + txt_size[0] + 1, y0 + int(1.5*txt_size[1])),
+            (x0 + txt_size[0] + 1, y0 + int(1.5 * txt_size[1])),
             txt_bk_color,
             -1
         )
         cv2.putText(img, text, (x0, y0 + txt_size[1]), font, 0.4, txt_color, thickness=1)
 
     return img
+
+
+def draw_distance(im, left, top, right, bottom, distance):
+    # 绘制bbox下沿中心坐标
+    y = int(bottom)
+    x = (left + right) // 2
+    cv2.circle(im, (x, y), 4, (255, 178, 50), thickness=-1)
+
+    # 绘制竖直线
+    cv2.line(im, (960, 0), (960, 1080), (0, 0, 255), thickness=1)
+
+    # 绘制distance
+    cv2.putText(im, distance + 'm', (left, bottom + 30), cv2.FONT_HERSHEY_SIMPLEX, 1, (255, 255, 0), 2)
+
+
+def draw_in_radar(camera_frame, radar_frame):
+    plt.cla()
+    plt.xlabel('x')
+    plt.ylabel('z')
+    plt.xlim(xmax=50, xmin=-50)
+    plt.ylim(ymax=300, ymin=0)
+    plt.title('Fusion In Radar Coordinate', fontsize='large', fontweight='bold', verticalalignment='center')
+    colors1 = '#00CED1'  # 点的颜色
+    colors2 = '#DC143C'
+    area = np.pi * 4 ** 2  # 点面积
+    if len(camera_frame) > 0:
+        plt.scatter(camera_frame[:, 0], camera_frame[:, 2], s=area, c=colors2, alpha=0.4, label='camera')
+    if len(radar_frame) > 0:
+        plt.scatter(radar_frame[:, 0], radar_frame[:, 2], s=area, c=colors1, alpha=0.4, label='radar')
+    plt.legend()
+    plt.pause(0.0001)
 
 
 _COLORS = np.array(
